move Visitor to a separate structure
